Updated Readme.md and fixed tokeninvalid bugs
