+++ Events/interactionCreate.js
@@ -1,5 +1,6 @@
 const Discord = require("discord.js")
 const { executeQuery } = require("../Fonctions/databaseConnect.js") 
+const { panelrole, ticketcategory } = require("../Config/TicketConfig.js")
 
 module.exports = async (bot, interaction, message) => {
 
@@ -27,6 +28,109 @@ module.exports = async (bot, interaction, message) => {
             await interaction.respond(entry === "" ? sortie.map(c => ({name: c, value: c})) : sortie.map(c => ({name: c, value: c})))
         }
     }
+
+    if(interaction.customId === 'ticketpannel') {
+
+        const openTicketEmbed = new Discord.EmbedBuilder()
+            .setTitle('Ticket en cours de création.')
+            .setDescription(`Veuillez patienter...`)
+            .setColor(bot.color)
+            .setTimestamp()
+
+        const msg = await interaction.reply({ embeds: [openTicketEmbed], ephemeral: true });
+            setTimeout(async() => {
+                const channelTicket = await interaction.guild.channels.create({
+                    name: `ticket-${interaction.user.username}`,
+                    type: 0,
+                    parent: ticketcategory,
+
+                    permissionOverwrites: [
+                        {
+                            id: interaction.user.id,
+                            allow: [
+                                Discord.PermissionFlagsBits.ViewChannel,
+                                Discord.PermissionFlagsBits.SendMessages,
+                                Discord.PermissionFlagsBits.ReadMessageHistory,
+                                Discord.PermissionFlagsBits.EmbedLinks,
+                                Discord.PermissionFlagsBits.AttachFiles,
+                                Discord.PermissionFlagsBits.UseExternalEmojis,
+                                Discord.PermissionFlagsBits.AddReactions,
+                            ],
+                        },
+                        {
+                            id: interaction.guild.roles.cache.get(panelrole).id,
+                            allow: [
+                                Discord.PermissionFlagsBits.ViewChannel,
+                                Discord.PermissionFlagsBits.SendMessages,
+                                Discord.PermissionFlagsBits.ReadMessageHistory,
+                                Discord.PermissionFlagsBits.EmbedLinks,
+                                Discord.PermissionFlagsBits.AttachFiles,
+                                Discord.PermissionFlagsBits.UseExternalEmojis,
+                                Discord.PermissionFlagsBits.AddReactions,
+                                Discord.PermissionFlagsBits.ManageMessages,
+                                Discord.PermissionFlagsBits.ManageGuildExpressions,
+                            ]
+                        },
+                        {
+                            id: interaction.guild.roles.everyone.id,
+                            deny: [
+                                Discord.PermissionFlagsBits.ViewChannel,
+                                Discord.PermissionFlagsBits.SendMessages,
+                                Discord.PermissionFlagsBits.ReadMessageHistory,
+                                Discord.PermissionFlagsBits.EmbedLinks,
+                                Discord.PermissionFlagsBits.AttachFiles,
+                                Discord.PermissionFlagsBits.UseExternalEmojis,
+                                Discord.PermissionFlagsBits.AddReactions,
+                            ],
+                        },
+                    ],
+                });
+                channelTicket.push({
+                    channelId: channel.id,
+                    userId: interaction.user.id,
+                    parentId: ticketcategory,
+                });
+                openTicketEmbed.setDescription(`Création du ticket...`);
+                msg.edit({ embeds: [openTicketEmbed], ephemeral: true });
+            }, 2000);
+
+            setTimeout(() => {
+                const storedChannelInfo = channelTicket[0];
+                const { channelId, parentId, userId } = storedChannelInfo;
+                const channel = interaction.bot.channels.cache.get(channelId);
+                const embedTicketOpen = new Discord.EmbedBuilder()
+                    .setTitle("Votre ticket est ouvert!")
+                    .setDescription(`Merci d'avoir ouvert ce ticket !\nPour avoir de l'aide, merci de patientez le temps qu'un membre du staff arrive !`)
+                    .setColor(bot.color)
+                    .setTimestamp()
+
+                const closeTicket = new Discord.ActionRowBuilder()
+                .addComponents(
+                    new Discord.ButtonBuilder()
+                        .setCustomId('closeTicket')
+                        .setLabel('Fermer le Ticket')
+                        .setStyle('Danger')
+                        .setEmoji('🔒'),
+                )
+                    
+                channel.send({ embeds: [embedTicketOpen], components: [closeTicket] });
+                openTicketEmbed.setDescription(`Envoie de l'embed dans le ticket...`);
+                msg.edit({ embeds: [openTicketEmbed], ephemeral: true });
+            }, 4000);
+
+            setTimeout(() => {
+                const storedChannelInfo = channelTicket[0];
+                const { channelId, parentId, userId } = storedChannelInfo;
+                const channel = interaction.bot.channels.cache.get(channelId);
+                channel.setTopic(`${userId}`)
+                openTicketEmbed.setDescription(`Votre ticket est prêt !\n<#${channelId}>`);
+                msg.edit({ embeds: [openTicketEmbed], ephemeral: true });
+                channel.send({ content: `<@${userId}>` }).then(msg => {
+                    msg.delete();
+                });
+            }, 6000);
+    };
+
 
     if (interaction.customId === "unwarn") {
         if(interaction.isButton()) {
